refactor: changed average stake calculation algorithm
Previous strategy would consider the time span the user remained allocated.

This was problematic because events payload contains only the block height,
not its timestamp.  The timestamp could only be retrieved by fetching the
entire block, which was a whole lot of data.

Even with caching, every month we would be working with brand new blocks,
so the script would take too much time to run and also cause spikes on
Infura API keys usage.

Now instead of considering the time span, the weight is the amount of
blocks the user remained staked. To make that work in terms of a month
time span like the usual, the trick is to find the first and the last
block mined within a given month.